+++ client/src/pages/studyCollectionsDetails/StudyUpdateModal.js
@@ -1,9 +1,30 @@
+/*
+ * Copyright 2019-2023 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 import React from "react";
 import PropTypes from "prop-types";
-import {Modal} from "react-bootstrap";
+import {Button, Modal} from "react-bootstrap";
 import StudyInputs from "../../common/forms/StudyInputs";
 
 const StudyUpdateModal = ({studies, isOpen, closeModal, handleUpdate}) => {
+
+  const handleStudyUpdate = (updatedStudies) => {
+
+  }
+
   return (
       <Modal
           show={isOpen}
@@ -16,6 +37,9 @@ const StudyUpdateModal = ({studies, isOpen, closeModal, handleUpdate}) => {
         <Modal.Body>
           <StudyInputs studies={studies} onChange={handleUpdate} />
         </Modal.Body>
+        <Modal.Footer>
+          <Button variant={"secondary"} onClick={closeModal}>Close</Button>
+        </Modal.Footer>
       </Modal>
   );
 }